requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
schedule>=1.2.0
loguru>=0.7.0
py-clob-client>=0.17.0
